Updating adsolver
Adding an emulate method to ADSolver.
The emulate method is essentially the same as the solve method (probably a lot of code duplication).
The difference is that for the emulators we take the source term to be a function of the source term and spectral width in addition to the wind, S=S(u,sf,cw). So the source term requires an augmented input.